layer: split unit tests
